Changing files location and installing Native Stack dependacy
