+++ App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {HomeScreen} from './src/home/HomeScreen';
+import {HomeScreen} from './src/pages/home/HomeScreen';
 
 const App = () => {
   return <HomeScreen />;
