+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@react-navigation/native": "^6.1.4",
+    "@react-navigation/native-stack": "^6.9.10",
     "react": "18.2.0",
     "react-native": "0.71.1",
     "react-native-safe-area-context": "^4.5.0",
