+++ src/pages/home/HomeScreen.tsx
@@ -1,6 +1,6 @@
 import {StyleSheet, View, FlatList} from 'react-native';
 import React from 'react';
-import {Header} from './components/Header';
+import {Header} from '../header/Header';
 import {HomeButton} from './components/HomeButton';
 import {HOME_SCREEN_BUTTONS_DATASET} from './components/HomeButtonsData';
 
@@ -13,7 +13,7 @@ export const HomeScreen = () => {
   return (
     <View>
       <Header
-        headerImagePath={require('./components/img/headerBackground.jpg')}
+        headerImagePath={require('../header/img/header_background.jpg')}
         headerText={HEADER_HOME_TEXT}
       />
       <View style={styles.listWrapper}>
