+++ src/pages/home/components/HomeButtonsData.tsx
@@ -1,0 +1,18 @@
+export const HOME_SCREEN_BUTTONS_DATASET = [
+  {
+    label: 'Encyclopédie',
+    imgPath: require('../img/encyclo_betes.png'),
+  },
+  {
+    label: 'Tracker',
+    imgPath: require('../img/tracker.png'),
+  },
+  {
+    label: 'Fossiles',
+    imgPath: require('../img/encyclo_fossiles.png'),
+  },
+  {
+    label: 'Mon Village',
+    imgPath: require('../img/my_village.png'),
+  },
+];
